+++ Ui/includes/gui/record_trigger_tcp_server.hh
@@ -1,11 +1,14 @@
 /*! \file record_trigger_tcp_server.hh
  *
- * \brief Lightweight TCP notification server for external record triggers.
+ * \brief Lightweight TCP server for external record triggers.
  */
 #pragma once
 
 #include <cstdint>
+#include <functional>
 
+#include <QByteArray>
+#include <QHash>
 #include <QTcpServer>
 #include <QTcpSocket>
 #include <QVector>
@@ -15,7 +18,7 @@ namespace holovibes::gui
 
 /*! \class RecordTriggerTcpServer
  *
- * \brief TCP server used to notify connected clients when a recording starts.
+ * \brief TCP server used to exchange recording start triggers with connected clients.
  */
 class RecordTriggerTcpServer
 {
@@ -23,6 +26,7 @@ class RecordTriggerTcpServer
     static RecordTriggerTcpServer& instance();
 
     void configure(bool enabled, quint16 port);
+    void set_record_start_callback(std::function<void()> callback);
     void notify_record_started();
 
     bool enabled() const { return enabled_; }
@@ -35,10 +39,13 @@ class RecordTriggerTcpServer
     void start();
     void stop();
     void accept_pending_connections();
+    void process_client_data(QTcpSocket* client);
     void remove_client(QTcpSocket* client);
 
     QTcpServer server_;
     QVector<QTcpSocket*> clients_;
+    QHash<QTcpSocket*, QByteArray> receive_buffers_;
+    std::function<void()> record_start_callback_;
     bool enabled_ = false;
     quint16 port_ = 50000;
 };